Revert "refactor: change vector3d node to use float"
This reverts commit eab08acd27f1bae22c7fe0ac9313380b0ff72c1d.

+++ common/vector3d.py
@@ -3,21 +3,21 @@ class VECTOR3D:
     def INPUT_TYPES(s):
         return {
             "required": {
-                "x": ("FLOAT", {
+                "x": ("INT", {
                     "default": 0, 
                     "min": -1024,
                     "max": 1024,
                     "step": 1,
                     "display": "number"
                 }),
-                "y": ("FLOAT", {
+                "y": ("INT", {
                     "default": 0, 
                     "min": -1024,
                     "max": 1024,
                     "step": 1,
                     "display": "number"
                 }),
-                "z": ("FLOAT", {
+                "z": ("INT", {
                     "default": 0, 
                     "min": -1024,
                     "max": 1024,
